// BermenMarch.MasterSideLetter.Web/ClientApp/src/app-common/models/provision.ts
export class Provision {
    id: number;
    provisionType: string;
    notes: string;
    content: string;
    fundInvestorId: number;
    fundName: string;
    investorName: string;
    fundSponsorName: string;
    sideLetterFileName: string;
    commitment: string;
    investorType: string;
    fundYear: string;
    createdDate: Date;
    modifiedDate: Date;
    htmlContent:string;
  }